Refactor of BackendHandler and restructure of BackgroundTask
